# backend/docs/swagger.json
{"swagger":"2.0","info":{"description":"This is the API documentation for the backend service.","title":"Backend API","contact":{},"version":"1.0"},"host":"localhost:8080","basePath":"/api","paths":{"/auth/login":{"post":{"description":"Authenticate a user and return a JWT token","consumes":["application/json"],"produces":["application/json"],"tags":["Auth"],"summary":"Login a user","parameters":[{"description":"Login request","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/handlers.LoginRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/handlers.AuthResponse"}},"400":{"description":"Invalid request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"401":{"description":"Invalid credentials","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"/auth/me":{"get":{"security":[{"BearerAuth":[]}],"description":"Verify the JWT token and return user information","consumes":["application/json"],"produces":["application/json"],"tags":["Auth"],"summary":"Verify authentication token","responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/handlers.AuthResponse"}},"401":{"description":"Invalid token","schema":{"type":"object","additionalProperties":{"type":"string"}}},"404":{"description":"User not found","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"/auth/signup":{"post":{"description":"Create a new user account","consumes":["application/json"],"produces":["application/json"],"tags":["Auth"],"summary":"Sign up a new user","parameters":[{"description":"Signup request","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/handlers.SignupRequest"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/handlers.AuthResponse"}},"400":{"description":"Invalid request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"409":{"description":"User already exists","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"/categories":{"get":{"description":"Retrieve a list of all available categories","produces":["application/json"],"tags":["Categories"],"summary":"Get all categories","responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/models.Category"}}},"500":{"description":"Failed to get categories","schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"post":{"security":[{"BearerAuth":[]}],"description":"Create a new category with the given name","consumes":["application/json"],"produces":["application/json"],"tags":["Categories"],"summary":"Create a new category","parameters":[{"description":"Category to create","name":"category","in":"body","required":true,"schema":{"$ref":"#/definitions/models.CategoryCreate"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/models.Category"}},"400":{"description":"Invalid request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Failed to create category","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"/categories/{id}":{"get":{"description":"Retrieve details of a specific category by ID","produces":["application/json"],"tags":["Categories"],"summary":"Get a specific category","parameters":[{"type":"string","description":"Category ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/models.Category"}},"400":{"description":"Invalid category ID","schema":{"type":"object","additionalProperties":{"type":"string"}}},"404":{"description":"Category not found","schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"put":{"security":[{"BearerAuth":[]}],"description":"Update an existing category's information","consumes":["application/json"],"produces":["application/json"],"tags":["Categories"],"summary":"Update a category","parameters":[{"type":"string","description":"Category ID","name":"id","in":"path","required":true},{"description":"Updated category","name":"category","in":"body","required":true,"schema":{"$ref":"#/definitions/models.CategoryUpdate"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/models.Category"}},"400":{"description":"Invalid request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Failed to update category","schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"delete":{"security":[{"BearerAuth":[]}],"description":"Delete an existing category","tags":["Categories"],"summary":"Delete a category","parameters":[{"type":"string","description":"Category ID","name":"id","in":"path","required":true}],"responses":{"204":{"description":"No Content"},"400":{"description":"Invalid category ID","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Failed to delete category","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"/categories/{id}/prompts":{"get":{"description":"Retrieve all prompts belonging to a specific category","produces":["application/json"],"tags":["Categories","Prompts"],"summary":"Get prompts in a category","parameters":[{"type":"string","description":"Category ID","name":"id","in":"path","required":true},{"type":"integer","description":"Page number for pagination","name":"page","in":"query"},{"type":"integer","description":"Number of items per page","name":"pageSize","in":"query"}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":true}},"400":{"description":"Invalid category ID","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Failed to get prompts","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"/health":{"get":{"description":"Get the current health status of the API","produces":["application/json"],"tags":["Health"],"summary":"Check API health","responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":true}}}}},"/prompts":{"get":{"description":"Retrieve a list of prompts with pagination and filtering options","produces":["application/json"],"tags":["Prompts"],"summary":"Get all prompts","parameters":[{"type":"integer","description":"Page number for pagination","name":"page","in":"query"},{"type":"integer","description":"Number of items per page","name":"pageSize","in":"query"},{"type":"string","description":"Filter by category ID","name":"categoryId","in":"query"},{"type":"string","description":"Search term for filtering prompts","name":"search","in":"query"}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":true}},"500":{"description":"Failed to get prompts","schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"post":{"security":[{"BearerAuth":[]}],"description":"Create a new prompt with the provided information","consumes":["application/json"],"produces":["application/json"],"tags":["Prompts"],"summary":"Create a new prompt","parameters":[{"description":"Prompt to create","name":"prompt","in":"body","required":true,"schema":{"$ref":"#/definitions/models.PromptCreate"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/models.Prompt"}},"400":{"description":"Invalid request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"401":{"description":"Not authenticated","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Failed to create prompt","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"/prompts/{id}":{"get":{"description":"Retrieve details of a specific prompt by ID","produces":["application/json"],"tags":["Prompts"],"summary":"Get a specific prompt","parameters":[{"type":"string","description":"Prompt ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/models.Prompt"}},"400":{"description":"Invalid prompt ID","schema":{"type":"object","additionalProperties":{"type":"string"}}},"404":{"description":"Prompt not found","schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"put":{"security":[{"BearerAuth":[]}],"description":"Update an existing prompt's information","consumes":["application/json"],"produces":["application/json"],"tags":["Prompts"],"summary":"Update a prompt","parameters":[{"type":"string","description":"Prompt ID","name":"id","in":"path","required":true},{"description":"Updated prompt information","name":"prompt","in":"body","required":true,"schema":{"$ref":"#/definitions/models.PromptUpdate"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/models.Prompt"}},"400":{"description":"Invalid request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"401":{"description":"Not authenticated","schema":{"type":"object","additionalProperties":{"type":"string"}}},"403":{"description":"Not authorized to update this prompt","schema":{"type":"object","additionalProperties":{"type":"string"}}},"404":{"description":"Prompt not found","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Failed to update prompt","schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"delete":{"security":[{"BearerAuth":[]}],"description":"Delete an existing prompt","tags":["Prompts"],"summary":"Delete a prompt","parameters":[{"type":"string","description":"Prompt ID","name":"id","in":"path","required":true}],"responses":{"204":{"description":"No Content"},"400":{"description":"Invalid prompt ID","schema":{"type":"object","additionalProperties":{"type":"string"}}},"401":{"description":"Not authenticated","schema":{"type":"object","additionalProperties":{"type":"string"}}},"403":{"description":"Not authorized to delete this prompt","schema":{"type":"object","additionalProperties":{"type":"string"}}},"404":{"description":"Prompt not found","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Failed to delete prompt","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"/prompts/{id}/like":{"post":{"security":[{"BearerAuth":[]}],"description":"Like or unlike a prompt for the authenticated user","produces":["application/json"],"tags":["Prompts"],"summary":"Toggle like on a prompt","parameters":[{"type":"string","description":"Prompt ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":true}},"400":{"description":"Invalid prompt ID","schema":{"type":"object","additionalProperties":{"type":"string"}}},"401":{"description":"Not authenticated","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Failed to process like","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}},"definitions":{"handlers.AuthResponse":{"type":"object","properties":{"error":{"type":"string"},"token":{"type":"string"},"user":{"$ref":"#/definitions/models.UserResponse"}}},"handlers.LoginRequest":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string"},"password":{"type":"string"}}},"handlers.SignupRequest":{"type":"object","required":["email","password","username"],"properties":{"email":{"type":"string"},"password":{"type":"string","minLength":6},"username":{"type":"string"}}},"models.Category":{"type":"object","properties":{"color":{"type":"string"},"created_at":{"type":"string"},"description":{"type":"string"},"icon_url":{"type":"string"},"id":{"type":"string"},"name":{"type":"string"},"prompt_count":{"type":"integer"},"updated_at":{"type":"string"}}},"models.CategoryCreate":{"type":"object","required":["name"],"properties":{"color":{"type":"string"},"description":{"type":"string"},"icon_url":{"type":"string"},"name":{"type":"string"}}},"models.CategoryUpdate":{"type":"object","properties":{"color":{"type":"string"},"description":{"type":"string"},"icon_url":{"type":"string"},"name":{"type":"string"}}},"models.Prompt":{"type":"object","properties":{"category_id":{"type":"string"},"content":{"type":"string"},"created_at":{"type":"string"},"description":{"type":"string"},"favorite_count":{"type":"integer"},"id":{"type":"string"},"image_url":{"type":"string"},"title":{"type":"string"},"updated_at":{"type":"string"},"user_id":{"type":"string"},"view_count":{"type":"integer"}}},"models.PromptCreate":{"type":"object","required":["category_id","content","title"],"properties":{"category_id":{"type":"string"},"content":{"type":"string"},"description":{"type":"string"},"image_url":{"type":"string"},"title":{"type":"string"}}},"models.PromptUpdate":{"type":"object","properties":{"category_id":{"type":"string"},"content":{"type":"string"},"description":{"type":"string"},"image_url":{"type":"string"},"title":{"type":"string"}}},"models.UserResponse":{"type":"object","properties":{"created_at":{"type":"string"},"email":{"type":"string"},"favoriteCount":{"description":"Ignored by GORM","type":"integer"},"id":{"type":"string"},"is_admin":{"type":"boolean"},"last_login":{"type":"string"},"promptCount":{"description":"Ignored by GORM","type":"integer"},"updated_at":{"type":"string"},"username":{"type":"string"}}}},"securityDefinitions":{"BearerAuth":{"type":"apiKey","name":"Authorization","in":"header"}}}